Extract JwtIssuerConfig into separate class
